Rename module and classes
