+++ ComposeGallery/src/androidTest/java/com/microsoft/device/display/samples/composegallery/TopAppBarTest.kt
@@ -22,8 +22,8 @@ import com.microsoft.device.display.samples.composegallery.ui.view.ComposeGaller
 import com.microsoft.device.display.samples.composegallery.ui.view.DetailPane
 import com.microsoft.device.display.samples.composegallery.ui.view.ListPane
 import com.microsoft.device.dualscreen.testutils.getString
-import com.microsoft.device.dualscreen.window_info.WindowInfo
-import com.microsoft.device.dualscreen.window_info.WindowSizeClass
+import com.microsoft.device.dualscreen.window.WindowSizeClass
+import com.microsoft.device.dualscreen.window.WindowState
 import org.junit.Rule
 import org.junit.Test
 
@@ -181,7 +181,7 @@ class TopAppBarTest {
         composeTestRule.setContent {
             ComposeGalleryTheme {
                 ComposeGalleryApp(
-                    WindowInfo(
+                    WindowState(
                         hasFold = false,
                         isFoldHorizontal = false,
                         foldBounds = Rect(),
