+++ ComposeGallery/src/main/java/com/microsoft/device/display/samples/composegallery/MainActivity.kt
@@ -10,20 +10,20 @@ import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
 import com.microsoft.device.display.samples.composegallery.ui.ComposeGalleryTheme
 import com.microsoft.device.display.samples.composegallery.ui.view.ComposeGalleryApp
-import com.microsoft.device.dualscreen.window_info.WindowInfo
-import com.microsoft.device.dualscreen.window_info.rememberWindowInfo
+import com.microsoft.device.dualscreen.window.WindowState
+import com.microsoft.device.dualscreen.window.rememberWindowState
 
 class MainActivity : AppCompatActivity() {
-    private lateinit var windowInfo: WindowInfo
+    private lateinit var windowState: WindowState
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         setContent {
-            windowInfo = rememberWindowInfo()
+            windowState = rememberWindowState()
 
             ComposeGalleryTheme {
-                ComposeGalleryApp(windowInfo)
+                ComposeGalleryApp(windowState)
             }
         }
     }
